requirements.txt
gspread
oauth2client
pandas
numpy
mip
datetime
icalendar
pytz

sendgrid
pybase64
python-dotenv
